Porting WMC insertion to geonetwork - develop
See https://github.com/georchestra/georchestra/issues/751 for the
details.

This PR proposes to add a webservice that allows WMC to be posted, which
will create a MD corresponding to the WMC received as input. Originally
developped for the geonetwork's fork of georchestra, this commit is a
port to the current develop branche.

Unit-test: missing the testImportWmc() to be ported back, for now it
fails on a NPE because of missing mocked objects.

TODO:
* finishing porting back the testsuite
* Needs live testing
* Needs specific extra developments (UI), because the whole interface
  from georchestra is built upon mapfishapp application.